Reworks the game API
The game API has been reworked to make the metadata more informative
and to simplify the game configuration.

To make the metadata more informative, the Parameter API has been
added. This API allows the description of the parameters associated
with each game.

The game configuration API has been simplified by using the parameter
information of the metadata to build the configuration instances.