Improved first playable version of VerbalFluency
The game now mixes the dices so they don't overlap with each other.

Time is now represented in seconds, making it more clean, in the top-center of the screen.

Added a dock in which the dice slots are situated. Changed the reset and check texts for dynamic buttons with still, hover and pressed sprites.