Removes demo code from project
The demo code has been removed from the project to create a new independent
project.

In addition, the source code has been removed to fix assets references in the
source code. Currently, when using this library in other project, assets must
be copied to the new project, due to the way that the assets were referenced in
the source code. This commit fixes this problem by using "import" statements
to reference the assets and by including the assets in the transpiled code.