Redesigns game configuration
Game configuration is now automatically generated from metadata using the
GameConfig.forMetadata static method. This method adds the properties defined
in the metadata to a GameConfig instance and adds a proxy that validates the
values set to this properties using the parameters defined in the metadata.

This will allow a faster development, as no configuration specific classes
are needed for each game.

In addition, the game callbacks have been moved to a class.
GameConfig instance that will