Adds GameBuilder to create games from game configurations
The GameBuilder class allows creating a game from a GameConfig. To do
so, the game's metadata has been added as an attribute to the GameConfig
class.

Ideally, the metadata classes should contain a reference to the related
game constructor. This way, the game builder could instance the game
from this information. However, due to a problem with Phaser in tests
this have not been done and the games should be registered manually in
the GameBuilder class. If the testing problem is fixed, this should be
changed.