Refactorization of the game structure
The game structure has been refactorized in order to separate the
data and logic from the presentation classes. In addition, a new
general game status and specific stage status classes have been
added.

Only the Verbal Fluency game has been refactorized and, currently,
the Recognition game is not working and it should be refactorized.