Adds new parameter that allows to choose dice face
A new parameter has been added to use as game metadata to allow to select
the dice face in some games. Also, this parameter was added to the games that
allows choose the dice face. By default this parameter is setting by random.
With this new parameter, some translations have been added.