Implements a new playback hearing game
A new game has been implemented based on Playback Hearing category. Added all
necessary files and modified the builders to include this new game.
Added a method in StageRenderer file to load the audio files to use in this new
game or any game that needs to use dice face sounds. Modified
StartStageRenderer to adjust the parameters box to show correctly all the games
configurations.