Modifies associated pairs game
The Associated Pairs game has been modified to make that if the correct dice
has not been selected, it will be hidden. When the correct dice is selected,
all the dices became visible again. Modifies the implementation to hide the
element that is visible in the result dock. Removes an unused function from
MainStageStatus.