Adds game input grid component
A game input grid component has been created to add a grid of inputs to the
games. With this component you can create a customized grid input and you can
modify the style. You must indicate the the number of input elements, the
element input size, the sprite to show over the element and the style.