Modifies Playback hearing game
The Playback Hearing game has been modified to change the response introduction
to use the new grid input component. With this change, the users can write all
the responses in the same view and watch if the introduced response is a valid
value. When the user complete all the inputs, must to check the responses and a
confirmation image will showed by each input element.